doc: usb: ci-hdrc-usb2: add property non-zero-ttctrl-ttha
If this property is not set, the max packet size is 1023 bytes, and if
the total of packet size for pervious transactions are more than 256 bytes,
it can't accept any transactions within this frame. The use case is single
transaction, but higher frame rate.

If this property is set, the max packet size is 188 bytes, it can handle
more transactions than above case, it can accept transactions until it
considers the left room size within frame is less than 188 bytes, software
needs to make sure it does not send more than 90%
maximum_periodic_data_per_frame. The use case is multiple transactions, but
less frame rate.

Signed-off-by: Peter Chen <peter.chen@nxp.com>

[ upstream commit: 4670ba6c9e6334d6d518d52499628f6bcb08fa95 ]