PCI: rcar: Add device tree support for r8a7793
Add "renesas,pcie-r8a7793" as a compatibility string for
"renesas,pcie-rcar-gen2".

This doesn't change the driver, so it does nothing by itself.  But it
does mean that checkpatch won't complain about a future patch that adds
"renesas,pci-r8a7793" to a DT, which helps ensure that shipped DTs use
documented compatibility strings.

[bhelgaas: changelog]
Signed-off-by: Simon Horman <horms+renesas@verge.net.au>
Signed-off-by: Bjorn Helgaas <bhelgaas@google.com>
Acked-by: Rob Herring <robh@kernel.org>
[ upstream commit: 0cf1337e0b83c16de4e7e98dad3a6afce6043fea ]